Fix clippy warnings to resolve CI/CD build failure
- Remove unnecessary cast in wf_common/src/common.rs
- Fix lifetime syntax in wf_crypto/src/wf_encryption_key.rs
- Replace ToString with Display implementation in wf_field/src/types.rs
- Fix redundant field names in wf_field/src/message_header.rs
- Update format strings to use inline variables throughout codebase
- Fix manual ok() implementation in wf_parser/src/header.rs
- Remove needless match expressions and question mark operators
- Add allow(dead_code) for unused struct fields

All 143 tests passing, clippy warnings resolved.

+++ wf_common/src/common.rs
@@ -202,7 +202,7 @@ pub fn concatinate_bits(
 
     /* Prepare byte arrays */
     let byte_array_2_shift = shift_right(byte_array_2, shift as isize);
-    let mut new_byte_array = vec![0; byte_length as usize];
+    let mut new_byte_array = vec![0; byte_length];
 
     /* Concatenation */
     let mut byte_cursor = 0;
